+++ packages/core/src/index.ts
@@ -3,7 +3,3 @@ export * from './ajax';
 export * from './observable';
 export * from './data';
 export * from './types';
-
-export const logInfo = (msg: string) => console.log(`[INFO] ${msg}`);
-export const logWarn = (msg: string) => console.warn(`[WARN] ${msg}`);
-export const logError = (msg: string) => console.trace(`[ERROR] ${msg}`);