+++ packages/ui/src/roxie-ui.ts
@@ -2,7 +2,7 @@ import 'jquery/dist/jquery.slim';
 import 'popper.js';
 import 'bootstrap';
 
-import $ from 'jquery/dist/jquery.slim';
+import $ from 'jquery';
 
 $(function() {
   $('[data-toggle="tooltip"]').tooltip();